refactor: moved table types
